requirements.txt
fastapi==0.70.0
uvicorn[standard]==0.15.0
websockets==14.2
websocket-client==1.8.0
pillow==11.1.0 
opencv-python==4.11.0.86
numpy==2.2.2